rag_vectordb/requirements.txt
flask
python-dotenv

langchain==0.1.16
langchain-openai==0.1.3
langchain-community==0.0.32

faiss-cpu
chromadb

unstructured[pdf,image]
pdfplumber
pytesseract

rank-bm25
numpy
requests
tqdm
